+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Added samples ([#218](https://github.com/opensearch-project/opensearch-php/pull/218))
 - Added support for PHP 8.3 and 8.4 ([#229](https://github.com/opensearch-project/opensearch-php/pull/229))
 - Added a Docker Compose config file for local development.
+- Added a test for the AWS signing client decorator
 ### Changed
 - Switched to PSR Interfaces
 - Increased PHP min version to 8.1